requirements.txt
python-binance==1.0.17
streamlit==1.33.0
